+++ examples/resources/graviteeam_org_form/resource.tf
@@ -1,0 +1,5 @@
+resource "graviteeam_org_form" "example" {
+  template = "LOGIN"
+  enabled  = true
+  content  = "<html><body><h1>Organization Login</h1></body></html>"
+}